# app/models/author.py
from app.models import Base
from sqlalchemy import Column, Integer, String, Text, Date, DateTime
from sqlalchemy.orm import relationship
from datetime import datetime, UTC


class Author(Base):
    __tablename__ = "authors"

    id = Column(Integer, primary_key=True)
    first_name = Column(String(50), nullable=False)
    last_name = Column(String(50), nullable=False)
    biography = Column(Text)
    birth_date = Column(Date)
    created_at = Column(DateTime, default=lambda: datetime.now(UTC))
    updated_at = Column(
        DateTime, default=lambda: datetime.now(UTC), onupdate=lambda: datetime.now(UTC)
    )

    # Relationships
    books = relationship("Book", back_populates="author")

    def __repr__(self):
        return f"<Author {self.first_name} {self.last_name}>"